requirements.txt
# Core Data Processing
pandas==2.2.0
numpy==1.26.4

# Data Fetching
yfinance==0.2.36
requests==2.31.0

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Dashboard
dash==2.16.1
dash-bootstrap-components==1.5.0
plotly==5.19.0

# Reporting
fpdf2==2.7.9
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
